Tweaking of the website template as per the Getting Started video
I have followed the video tutorial provided by Mark in order to add additional baseline functionality to the website

This has involved replacing the "secure" modules and relevant references with "sos" (sandwich ordering service), as well as the addition of the contacts table to the database. Ihave also managed to set-up data retrieval from the database which is then displayed on the website body.