Initial setup for stage 1 functionality
I have created a navigation bar for both the owner and the customer to access the relevant content. I have additionally worked out and setup the time based restriction on the end of the customer (such that they cannot access orders post 11 AM).

This has involved changes to the owner and user menu handlebars files along with the menu_route route module.