Finalised client side javascript functionality on the menu page (to incorporate the checkout)
This has involved the addition of some client side javascript functions within the user_menu handlebars file.

This has also involved the initial configuration of a checkout route, checkout_route.js, which receives the data submitted via the POST request.

Still to do: finish the checkout route functionality and create the orders module to handle CRUD operations to the database.