Finished checkout functionality
This has involved updates to the orders module where I have created an updateLast function to add additional data (such as the 10 digit order number) to the last inserted record.

Asides from this, I have updated the checkout handlebars page to display details of the user's order/s that are currently incomplete.