Completion of tests and functionality for Stage 2 Part 2
This has involved the usage of nodemailer and qrcode to send a message with a QRCode encoded with the user's 10 digit order number upon a user confirming their checked out order

There have been changes to the orders module and orders route to facilitate this functionality. I have also created an email config file storing the testing email-address and password (this is ignored by git).