Stage 2 part 3 functionality complete
This has involved changes to the orders route and module, adding the functionality to view a specific order in detail as well as embedding the QR Code on the page

I have additionally added an order history functionality to view orders that have been marked as delivered.

The main change has been through the getOrderById function in the order module which works in conjunction with the function in the order route, passing the ctx id parameter over in order to server the appropriate data.